+++ Cargo.toml
@@ -54,22 +54,6 @@ cfg-if = "1.0.0"
 sha2 = "0.10.7"
 proptest = "1.2.0"
 
-[[bench]]
-name = "recursive-snark"
-harness = false
-
-[[bench]]
-name = "compressed-snark"
-harness = false
-
-[[bench]]
-name = "compute-digest"
-harness = false
-
-[[bench]]
-name = "sha256"
-harness = false
-
 [features]
 default = []
 # Compiles in portable mode, w/o ISA extensions => binary can be executed on all systems.
