--- SUPPORT.md
@@ -1,10 +1,0 @@
-# Support
-
-## How to file issues and get help  
-
-This project uses GitHub Issues to track bugs and feature requests. Please search the existing issues before filing new issues to avoid duplicates.  For new issues, file your bug or 
-feature request as a new Issue.
-
-## Microsoft Support Policy  
-
-Support for this project is limited to the resources listed above.
