+++ src/compress_snark/math.rs
