+++ src/compress_snark/polys/eq.rs
