+++ src/gadgets/mod.rs
@@ -2,7 +2,8 @@
 pub(crate) mod cccs;
 pub mod ecc;
 pub(crate) mod ext_allocated_num;
+pub(crate) mod lcccs;
 pub(crate) mod nonnative;
-pub(crate) mod r1cs;
+pub(crate) mod r1cs_instance;
 pub(crate) mod sumcheck;
 pub mod utils;
