+++ src/nimfs/ccs/mod.rs
@@ -2,5 +2,3 @@ pub mod cccs;
 #[allow(clippy::module_inception)]
 pub mod ccs;
 pub mod lcccs;
-pub mod pedersen;
-pub mod util;
