+++ src/nimfs/mod.rs
@@ -4,8 +4,6 @@
 
 pub mod ccs;
 pub mod multifolding;
-pub mod pcd_aux_circuit;
-pub mod pcd_circuit;
 
 pub mod espresso;
 pub mod util;
